api/requirements.txt
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn==0.31.0
fastapi==0.115.0
orjson==3.9.7
pydantic==2.4.0
pydantic-settings==2.0.3
greenlet==2.0.2
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.1
asyncpg==0.28.0
psycopg2-binary==2.9.7
alembic==1.12.0
aiofiles==23.2.1
pathlib==1.0.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
aiosmtplib==2.0.2
jinja2==3.1.2
docker==6.1.3
aiohttp==3.10.9
scrapyd-api==0.0.4
telegraph==2.2.0
deep-translator==1.11.4
openai==1.55.0
beautifulsoup4==4.12.3
pytest==7.4.2
pytest-asyncio==0.21.1
flake8==6.1.0